requirements.txt
requests==2.28.1

yaml~=0.2.5
pyyaml~=6.0
pandas~=1.3.4
python-dotenv~=0.21.0